fix: 修复 multiple 参数同 customPinyin 或者 surname 同时使用的问题
